Move characterization to library folder
