+++ docs/apidocs/characterization.rst
@@ -1,6 +1,6 @@
 .. _qiskit-experiments-characterization:
 
-.. automodule:: qiskit_experiments.characterization
+.. automodule:: qiskit_experiments.library.characterization
    :no-members:
    :no-inherited-members:
    :no-special-members:
