+++ qiskit_experiments/library/__init__.py
@@ -63,8 +63,8 @@ Composite Experiments
 """
 
 from qiskit_experiments.quantum_volume import QuantumVolume
-from qiskit_experiments.characterization import T1, T2Ramsey, QubitSpectroscopy, EFSpectroscopy
 from qiskit_experiments.calibration import DragCal, Rabi, EFRabi
 from qiskit_experiments.composite import ParallelExperiment, BatchExperiment
+from .characterization import T1, T2Ramsey, QubitSpectroscopy, EFSpectroscopy
 from .randomized_benchmarking import StandardRB, InterleavedRB
 from .tomography import StateTomography, ProcessTomography
